pyproject.toml
[tool.poetry]
name = "fastsecond"
version = "0.1.0"
description = ""
authors = ["yaseen1998 <yasseenkouthe@eng.hu.edu.jo>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.78.0"
SQLAlchemy = "^1.4.39"
uvicorn = "^0.18.2"
databases = "^0.6.0"
asyncpg = "^0.25.0"
psycopg2-binary = "^2.9.3"
aiosqlite = "^0.17.0"
alembic = "^1.8.0"
python-decouple = "^3.6"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
